+++ unidock/unidock_engine/src/task/DockTask.h
@@ -70,7 +70,7 @@ private:
     // CPU
     int n_atom_all_flex = 0;
     int n_dihe_all_flex = 0;
-    std::vector<int> list_i_real; // record indices/ranges of coords & dihedrals in Real data, size: nflex * exhaustiveness
+    std::vector<size_t> list_i_real; // record indices/ranges of coords & dihedrals in Real data, size: nflex * exhaustiveness
     FlexPose* flex_pose_list_res; // size: nflex * exhaustiveness todo: add a member score instead of these usurpers
     // Finally, use center[3] to record intra, inter, penalty
     // then use orientation[4] to record Predicted Free Energy of Binding, Total score, inter(contains penalty) score, conf_independent part
